Issue #1859: Refactor MediaSession.observePlaybackState()
